feat : 관리자 피드 리스트 검색 조회 기능 추가

+++ src/main/java/com/example/demo/src/user/entity/specification/UserSpecification.java
@@ -4,7 +4,6 @@ import com.example.demo.common.entity.BaseEntity;
 import com.example.demo.src.user.entity.User;
 import org.springframework.data.jpa.domain.Specification;
 
-import javax.swing.plaf.nimbus.State;
 import java.time.LocalDateTime;
 
 public class UserSpecification {
@@ -12,7 +11,7 @@ public class UserSpecification {
         return ((root, query, criteriaBuilder) ->
                 criteriaBuilder.like(root.get("name"), name));
     }
-    public static Specification<User> equalsId(Integer id){
+    public static Specification<User> equalsId(Long id){
         return ((root, query, criteriaBuilder) ->
                 criteriaBuilder.equal(root.get("id"), id));
     }
